requirements.txt
tensorflow==1.13.1
psycopg2-binary==2.7.5
Pillow==5.2.0
numpy==1.15.4
requests==2.21.0
tqdm==4.23.0
hydrosdk==2.0.0rc9
hydro-serving-grpc==2.0.3
hs==2.1.0rc6
namesgenerator==0.3
kfp==0.1.19
scikit-learn==0.20.2
mlflow==1.0.0
python-decouple==3.1